Stats window to a separated file
